refactor: separate container and presentation components
- Application → Application (container) + ApplicationView (presentation)
- YouTubePlayer → YouTubePlayer (container) + VideoPlayerView (presentation)
- Stories now use pure presentation components with mock data
- Remove Root.stories.tsx and YouTubePlayer.stories.tsx (container stories)
- Add ApplicationView.stories.tsx (Default, Loading, Error, WithSelectedSong)
- Add VideoPlayerView.stories.tsx (Default, Playing, Loading, Error, NoResults, SingleVideo)
