--- applications/web/sources/components/Root.stories.tsx
@@ -1,17 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import { Root } from './Root.tsx';
-
-const meta = {
-  title: 'Components/Root',
-  component: Root,
-  parameters: {
-    layout: 'fullscreen',
-  },
-} satisfies Meta<typeof Root>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {};
